requirements.txt
pandas
flask
python-decouple
pymysql
flask_sqlalchemy
flask_migrate
python-decouple
